appengine/requirements.txt
google-cloud-bigquery==1.25.0
Flask==1.1.2